README.md
# RecipeCookBook
Simple cooking app with recipe view, search and insert.
The idea of this application was to create an app that will be used as help in cooking recipes with ingredients that users already possess.  Application named “Kuharica” allows login and user registration, recipe review and input including recipe searches using recipe name or ingredient. This is my first android application made for croatian community. Cheers